Fix build and fetchWorker import
